Renamed libcxx to libncxx abd removed libc (#5)
