--- libs/libc/include_with_prefix/assert.h
@@ -1,6 +1,0 @@
-#pragma once 
-
-#define assert(condition) \
-do { \
- if (!(condition)) __builtin_trap(); \
-} while(false)
