--- libs/libc/include_with_prefix/nos_string.h
@@ -1,16 +1,0 @@
-#pragma once
-
-#include <stddef.h>
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-int nos_memcmp(const void* ptr1, const void* ptr2, size_t size);
-void* nos_memcpy(void* destination, const void* source, size_t size);
-void* nos_memmove(void* destination, const void* source, size_t size);
-void* nos_memset(void* ptr, int c, size_t size);
-
-#ifdef __cplusplus
-}
-#endif
